Refactor play-game function to be pure
Added arguments for the board and the player sequence to make the play-game a
pure function (uses only arguments passed to it).  There are still some side
effects however that is only in printing out to the standard out.

Using arguments allows you to start the game with different boards and a
different player sequence.  This can be useful for testing purposes too.